Adding kwargs to dump_yaml
This is required to add extra arguments such as the line width. Unity internally does not split as soon as the default python yaml parser does. We have some files in VC that we modify with our tools and without the proper line width this makes our local repos dirty each time we apply some tooling.